Project_flipkart_extract_data/requirement.txt
Flask
python-dotenv
pymongo
bs4
requests
geocoder
pandas
matplotlib
seaborn